Model makes much more sense now
The model now generates outputs in a proper range (if given at least 2
epochs). [Float, float, float... float] for 8 values. The first two are
the X, Y axis of the analog stick, those can be kept mostly as is (with
anything < 0 being set to 0, anything > 1 set to 1).

The rest are confidence intervals in button presses. If above a set
number (we'll say .5 for now), we press that button. Otherwise, don't
press/send release on the other buttons. Since we don't have memory,
assume we release every button we don't press.